Keep typed IRIs out of completion entries
Rather than add everything in PropertyTypes to the type directly, only
add those keys (and any IRIs) when they already appear in the object
being checked. Thus, typed properties will be typed, but will not
contribute to the completions.

Previously, each property in PropertyTypes would appear as a completion
entry—and worse, because they had to be quoted and thus the completions
began with a double-quote character, they'd all appear *first* in the
completion list.